refactor: reorganize js files as modules for a cleaner index.js
