+++ library/index.html
@@ -71,20 +71,15 @@
               </div>
 
               <div class="carousel-pagination">
-                <div class="carousel-pagination__button" role="button" aria-current="true">
-                </div>
+                <div class="carousel-pagination__button" role="button" aria-current="true"></div>
 
-                <div class="carousel-pagination__button" role="button">
-                </div>
+                <div class="carousel-pagination__button" role="button"></div>
 
-                <div class="carousel-pagination__button" role="button">
-                </div>
+                <div class="carousel-pagination__button" role="button"></div>
 
-                <div class="carousel-pagination__button" role="button">
-                </div>
+                <div class="carousel-pagination__button" role="button"></div>
 
-                <div class="carousel-pagination__button" role="button">
-                </div>
+                <div class="carousel-pagination__button" role="button"></div>
               </div>
 
             </div>
@@ -417,8 +412,7 @@
 
   </div>
 
-  <script src="./scripts/index.js"></script>
-  <script src="./scripts/self-check.js"></script>
+  <script type="module" src="./scripts/index.js"></script>
 </body>
 
 </html>
